feat: :recycle: moved map and removed list route
