+++ app/components/navigation-bar.ts
@@ -1,27 +1,27 @@
-import Component from "@glimmer/component";
 import RouterService from "@ember/routing/router-service";
 import { service } from "@ember/service";
+import Component from "@glimmer/component";
 
 interface Link {
-    route: string;
-    label: string;
+  route: string;
+  label: string;
 }
 
 export default class NavigationBar extends Component {
-    @service declare router: RouterService;
+  @service declare router: RouterService;
 
-    links: Array<Link> = [
-        {
-            route: "home.list",
-            label: "Agendapunten",
-        },
-        {
-            route: "home.map",
-            label: "Kaart",
-        },
-        {
-            route: "sessions",
-            label: "Zittingen",
-        },
-    ];
+  links: Array<Link> = [
+    {
+      route: "home",
+      label: "Agendapunten",
+    },
+    {
+      route: "map",
+      label: "Kaart",
+    },
+    {
+      route: "sessions",
+      label: "Zittingen",
+    },
+  ];
 }
