+++ app/router.js
@@ -10,10 +10,8 @@ Router.map(function () {
   this.route("detail", { path: "/agendapunt/:id" });
   this.route("municipality", { path: "/gemeente/:municipality" });
   this.route("help", { path: "/help" });
-  this.route("home", { path: "/" }, function (req, res) {
-    this.route("list", { path: "/" });
-    this.route("map", { path: "/kaart" });
-  });
+  this.route("home", { path: "/" });
+  this.route("map", { path: "/kaart" });
   this.route("sessions", { path: "/zittingen" }, function () {
     this.route("session", { path: "/:session_id" });
   });
