--- app/templates/home/list.hbs
@@ -1,26 +1,0 @@
-<section
-  {{did-insert this.setupInViewport}}
-  class="au-o-flow au-u-flex au-u-flex--column"
->
-  {{#if @model}}
-    {{#each @model as |item|}}
-      <DecisionCard
-        @startDate={{item.session.startedAt}}
-        @endDate={{item.session.endedAt}}
-        @item={{item}}
-      />
-    {{/each}}
-  {{else}}
-    <AuAlert @title={{"Oeps!"}} @skin={{"warning"}} @icon={{"archive"}}>
-      <p>Deze gemeente heeft spijtig genoeg geen resultaten!</p>
-    </AuAlert>
-  {{/if}}
-</section>
-{{#if @model}}
-  <div
-    id="loader"
-    style="height: 80px; display: flex; justify-content: center; align-items: center;"
-  >
-    <AuLoader />
-  </div>
-{{/if}}
