+++ app/templates/map.hbs
